fix save report button issue with no watchtower table
