+++ client/src/components/Dashboard.tsx
@@ -7,21 +7,23 @@ import {
   Grid,
   Paper,
 } from '@mui/material';
-
 import Layout from './Layout';
 import StatusBox from './StatusBox';
 import DataStats from './DataStats';
 import GraphContainer from './GraphContainer';
 import BedrockAnalysis from './BedrockAnalysis';
 import { ProvisionFormData, Metrics } from '../../types/types';
+import Swal from 'sweetalert2';
 
 const defaultTheme = createTheme();
 
-export default function Dashboard(): JSX.Element | null {
+const Dashboard = (): JSX.Element => {
   const [currentProvision, setCurrentProvision] =
     useState<ProvisionFormData | null>(null);
   const [currentMetrics, setCurrentMetrics] = useState<Metrics | null>(null);
   const [runGraph, setRunGraph] = useState<boolean>(false);
+  const [table, setTable] = useState<string[] | null>([]);
+  const [save, setSave] = useState<boolean>(false);
 
   const handleFormSubmit = async (data: ProvisionFormData): Promise<void> => {
     try {
@@ -63,6 +65,55 @@ export default function Dashboard(): JSX.Element | null {
     fetchMetrics();
   }, [currentProvision]);
 
+  useEffect(() => {
+    const getTables = async (): Promise<void> => {
+      const response: Response = await fetch(`/api/tables`);
+
+      if (!response.ok) {
+        throw new Error(
+          `HTTP error status:${response.status}, AWS security token is incorrect!`
+        );
+      }
+
+      const data = await response.json();
+      if (!data.includes('WatchTowerUserProfiles')) {
+        Swal.fire({
+          title: 'Are you sure?',
+          text: 'AWS Account Info Submitted! We will create a table "WatchTowerUserProfiles" to save reports.',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#70c0c2',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes!',
+        }).then(async (result) => {
+          if (result.isConfirmed) {
+            const response: Response = await fetch('/api/createTable');
+            const data = await response.json();
+            Swal.fire({
+              title: 'Table created!',
+              text: `Table "WatchTowerUserProfiles1" is created successfully in your DynamoDB!`,
+              icon: 'success',
+              confirmButtonColor: '#70c0c2',
+            });
+            setTable(data);
+            setSave(true);
+          }
+        });
+      }
+      setTable(data);
+    };
+
+    getTables().catch((err) => {
+      console.error("Couldn't get tables' name", err);
+      Swal.fire({
+        title: 'Oops...',
+        text: err.message,
+        icon: 'error',
+        confirmButtonColor: '#70c0c2',
+      });
+    });
+  }, []);
+
   return (
     <ThemeProvider theme={defaultTheme}>
       <CssBaseline />
@@ -86,7 +137,7 @@ export default function Dashboard(): JSX.Element | null {
                     currentMetrics={currentMetrics}
                   /> // Render DataStats when currentProvision is set
                 ) : (
-                  <StatusBox onSubmit={handleFormSubmit} /> // Render StatusBox when currentProvision is null
+                  <StatusBox table={table} onSubmit={handleFormSubmit} /> // Render StatusBox when currentProvision is null
                 )}
               </Paper>
             </Grid>
@@ -113,6 +164,7 @@ export default function Dashboard(): JSX.Element | null {
                   currentProvision={currentProvision}
                   currentMetrics={currentMetrics}
                   runGraph={runGraph}
+                  save={save}
                 />
               </Paper>
             </Grid>
@@ -121,4 +173,6 @@ export default function Dashboard(): JSX.Element | null {
       </Layout>
     </ThemeProvider>
   );
-}
+};
+
+export default Dashboard;
