+++ client/src/components/StatusBox.tsx
@@ -14,12 +14,10 @@ import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { format } from 'date-fns';
 import { StatusBoxProps } from '../../types/types';
-import Swal from 'sweetalert2';
 
-const StatusBox = ({ onSubmit }: StatusBoxProps): JSX.Element => {
+const StatusBox = ({ table, onSubmit }: StatusBoxProps): JSX.Element => {
   const [startDate, setStartDate] = useState<Date | null>(null);
   const [endDate, setEndDate] = useState<Date | null>(null);
-  const [table, setTable] = useState<string[] | null>(null);
   const [tableName, setTableName] = useState<string>('');
 
   const handleSubmit = (e: React.FormEvent): void => {
@@ -33,56 +31,6 @@ const StatusBox = ({ onSubmit }: StatusBoxProps): JSX.Element => {
       endTime,
     });
   };
-
-  useEffect(() => {
-    const getTables = async (): Promise<void> => {
-      const response: Response = await fetch(`/api/tables`);
-
-      if (!response.ok) {
-        console.log(response);
-        throw new Error(
-          `HTTP error status:${response.status}, AWS security token is incorrect!`
-        );
-      }
-
-      const data = await response.json();
-      if(!data.includes('WatchTowerUserProfiles')){
-        Swal.fire({
-          title: 'Are you sure?',
-          text: 'AWS Account Info Submitted! We will create a table "WatchTowerUserProfiles" to save reports.',
-          icon: 'warning',
-          showCancelButton: true,
-          confirmButtonColor: '#70c0c2',
-          cancelButtonColor: '#d33',
-          confirmButtonText: 'Yes!',
-        }).then(async (result) => {
-          if (result.isConfirmed) {
-            const response: Response = await fetch('/api/createTable');
-            const data = await response.json();
-            Swal.fire({
-              title: 'Table created!',
-              text: `Table "WatchTowerUserProfiles" is created successfully in your DynamoDB!`,
-              icon: 'success',
-              confirmButtonColor: '#70c0c2',
-            });
-            setTable(data);
-          }
-
-        });
-      }
-      setTable(data);
-    };
-
-    getTables().catch((err) => {
-      console.error("Couldn't get tables' name", err);
-      Swal.fire({
-        title: 'Oops...',
-        text: err.message,
-        icon: 'error',
-        confirmButtonColor: '#70c0c2',
-      });
-    });
-  }, []);
 
   return (
     <Container maxWidth='sm' style={{ overflow: 'visible' }}>
